+++ js/test/date/testdatefmt_bs_Cyrl_BA.js
@@ -1,7 +1,7 @@
 /*
  * testdatefmt_bs_Cyrl_BA.js - test the date formatter object in Cyrillic Bosnian
  *
- * Copyright © 2012-2015,2017-2018, JEDLSoft
+ * Copyright © 2012-2015,2017-2018,2020 JEDLSoft
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -540,7 +540,7 @@ module.exports.testdatefmt_bs_Cyrl_BA = {
             second: 0,
             millisecond: 0
         });
-        test.equal(fmt.format(date), "9");
+        test.equal(fmt.format(date), "с");
         test.done();
     },