+++ js/test/daterange/testdatefmtrange_vi_VN.js
@@ -608,7 +608,7 @@ module.exports.testdatefmtrange_vi_VN = {
             second: 0,
             millisecond: 0
         });
-        test.equal(fmt.format(start, end), "Thg 11, 2011 – Thg 1, 2014");
+        test.equal(fmt.format(start, end), "thg 11, 2011 – thg 1, 2014");
         test.done();
     },
     testDateRngFmtVNRangeMultiYearLong: function(test) {
@@ -634,7 +634,7 @@ module.exports.testdatefmtrange_vi_VN = {
             second: 0,
             millisecond: 0
         });
-        test.equal(fmt.format(start, end), "Tháng 11, 2011 – Tháng 1, 2014");
+        test.equal(fmt.format(start, end), "tháng 11, 2011 – tháng 1, 2014");
         test.done();
     },
     testDateRngFmtVNRangeMultiYearFull: function(test) {
@@ -660,7 +660,7 @@ module.exports.testdatefmtrange_vi_VN = {
             second: 0,
             millisecond: 0
         });
-        test.equal(fmt.format(start, end), "Tháng 11, 2011 – Tháng 1, 2014");
+        test.equal(fmt.format(start, end), "tháng 11, 2011 – tháng 1, 2014");
         test.done();
     },
     testDateRngFmtVNManyYearsFull: function(test) {