+++ tools/cldr/genclockprefs.js
@@ -21,7 +21,6 @@
  * This code is intended to be run under node.js
  */
 var fs = require('fs');
-var cldr = require("cldr-data");
 
 var common = require('./common');
 var merge = common.merge;
@@ -39,16 +38,15 @@ function usage() {
     "  the top level of the ilib locale data directory");
     process.exit(1);
 }
-var cldrDirName;
 var localeDirName;
-process.argv.forEach(function (val, index, array) {
+process.argv.forEach(function (val) {
     if (val === "-h" || val === "--help") {
         usage();
     }
 });
 localeDirName = process.argv[2] || "tmp";
 
-var locales = cldr.availableLocales;
+var locales = require("cldr-core/availableLocales.json").availableLocales.full;
 
 console.log("genclockprefs - generate clock preferences information files.\n" +
 "Copyright (c) 2013-2018 LGE\n");
@@ -57,7 +55,6 @@ console.log("locale dir: " + localeDirName);
 if (!fs.existsSync(localeDirName)) {
     common.makeDirs(localeDirName);
 }
-var languageData;
 
 function loadFile(pathname) {
     var ret = undefined;
@@ -165,24 +162,25 @@ function getClockPrefs(locale) {
     // always read the gregorian calendar settings to make it easy, because none of the locales use different
     // time formats depending on the calendar
     var greg, timeformat;
-    if (locale.substring(0,2) === "ku") {
-        greg = require(path.join("cldr-data/main", "ckb", "ca-gregorian.json"));
-        timeformat = greg.main["ckb"].dates.calendars.gregorian.timeFormats.short;
-    } else {
-        greg = require(path.join("cldr-data/main", locale, "ca-gregorian.json"));
+    locale = (locale.substring(0,2) === "ku") ? "ckb" : locale;
+    var filename = path.join("cldr-dates-full/main", locale, "ca-gregorian.json");
+    try {
+        greg = require(filename);
         timeformat = greg.main[locale].dates.calendars.gregorian.timeFormats.short;
-    }
 
-    if (timeformat.indexOf("H") != -1) {
-        clockprefs["clock"] = "24";
-    } else if (timeformat.indexOf("h") != -1) {
-        clockprefs["clock"] = "12";
-    } else {
-        console.log("could not find default clock preference for locale " + locale);
+        if (timeformat.indexOf("H") != -1) {
+            clockprefs["clock"] = "24";
+        } else if (timeformat.indexOf("h") != -1) {
+            clockprefs["clock"] = "12";
+        } else {
+            console.log("could not find default clock preference for locale " + locale);
+        }
+        //console.log("time format is :"+JSON.stringify(timeformat));
+        //console.log("clock preference is :"+JSON.stringify(clockprefs));
+        return clockprefs;
+    } catch (e) {
+        console.log("Could not find file " + filename + " ... skipping.");
     }
-    //console.log("time format is :"+JSON.stringify(timeformat));
-    //console.log("clock preference is :"+JSON.stringify(clockprefs));
-    return clockprefs;
 }
 
 var language, region, script, files;
@@ -190,12 +188,14 @@ var language, region, script, files;
 console.log("Reading locale data into memory...");
 for (var i = 0; i < locales.length; i++) {
     var pref = getClockPrefs(locales[i]);
-    if (locales[i] === "root") {
-        // special case because "root" is not a valid locale specifier#
-        localeData.data = pref;
-    } else {
-        var l = new Locale(locales[i]);
-        setLocaleData(pref, l);
+    if (pref) {
+        if (locales[i] === "root") {
+            // special case because "root" is not a valid locale specifier#
+            localeData.data = pref;
+        } else {
+            var l = new Locale(locales[i]);
+            setLocaleData(pref, l);
+        }
     }
 }