+++ tools/cldr/genweekdata.js
@@ -20,11 +20,9 @@
  * This code is intended to be run under node.js
  */
 var fs = require('fs');
-var supplementalData = require("cldr-data/supplemental/weekData.json");
+var supplementalData = require("cldr-core/supplemental/weekData.json");
 
-var unifile = require('./unifile.js');
 var common = require('./common.js');
-var UnicodeFile = unifile.UnicodeFile;
 var coelesce = common.coelesce;
 var mkdirs = common.makeDirs;
 var path = require("path");